+++ src/machine/leinSupport.ts
@@ -1,11 +1,18 @@
+import {
+    logger,
+} from "@atomist/automation-client";
+import { GitHubRepoRef } from "@atomist/automation-client/operations/common/GitHubRepoRef";
 import { GitProject } from "@atomist/automation-client/project/git/GitProject";
 import * as clj from "@atomist/clj-editors";
 import {
+    allSatisfied,
+    Builder,
     editorAutofixRegistration,
     ExecuteGoalResult,
     ExtensionPack,
+    hasFile,
+    ProjectLoader,
     RunWithLogContext,
-    SoftwareDeliveryMachine,
 } from "@atomist/sdm";
 import * as build from "@atomist/sdm/dsl/buildDsl";
 import {
@@ -13,18 +20,21 @@ import {
     VersionGoal,
 } from "@atomist/sdm/goal/common/commonGoals";
 import { branchFromCommit } from "@atomist/sdm/internal/delivery/build/executeBuild";
-import { leinBuilder } from "@atomist/sdm/internal/delivery/build/local/lein/leinBuilder";
 import {
     executeVersioner,
     ProjectVersioner,
 } from "@atomist/sdm/internal/delivery/build/local/projectVersioner";
+import { SpawnBuilder } from "@atomist/sdm/internal/delivery/build/local/SpawnBuilder";
 import { IsLein } from "@atomist/sdm/mapping/pushtest/jvm/jvmPushTests";
 import {
     DefaultDockerImageNameCreator,
     DockerOptions,
     executeDockerBuild,
 } from "@atomist/sdm/pack/docker/executeDockerBuild";
-import { spawnAndWatch } from "@atomist/sdm/util/misc/spawned";
+import {
+    asSpawnCommand,
+    spawnAndWatch,
+} from "@atomist/sdm/util/misc/spawned";
 import * as df from "dateformat";
 import * as path from "path";
 
@@ -34,11 +44,10 @@ export const LeinSupport: ExtensionPack = {
     version: "0.1.0",
     configure: sdm => {
 
-        // TODO cd atomist.sh builder
         sdm.addBuildRules(
             build.when(IsLein)
                 .itMeans("Lein build")
-                .set(leinBuilder(sdm.configuration.sdm.projectLoader, "lein do clean, dynamodb-local test")),
+                .set(leinBuilder(sdm.configuration.sdm.projectLoader)),
         );
 
         sdm.addGoalImplementation("leinVersioner", VersionGoal,
@@ -51,24 +60,61 @@ export const LeinSupport: ExtensionPack = {
                     {
                         ...sdm.configuration.sdm.docker.jfrog as DockerOptions,
                         dockerfileFinder: async () => "docker/Dockerfile",
-                    }), { pushTest: IsLein })
+                    }), { pushTest: allSatisfied(IsLein, hasFile("docker/Dockerfile")) })
             .addAutofixes(
                 editorAutofixRegistration(
-                    {name: "cljformat",
+                    {
+                        name: "cljformat",
                         editor: async p => {
                             await clj.cljfmt(p.baseDir);
                             return p;
                         },
+                        pushTest: IsLein,
                     }));
 
     },
 };
 
+function leinBuilder(projectLoader: ProjectLoader): Builder {
+    return new SpawnBuilder(
+        {
+            projectLoader,
+            options: {
+                name: "atomist.sh",
+                commands: [asSpawnCommand("./atomist.sh", {})],
+                errorFinder: (code, signal, l) => {
+                    return code !== 0;
+                },
+                logInterpreter: log => {
+                    return {
+                        // We don't yet know how to interpret clojure logs
+                        relevantPart: undefined,
+                        message: "lein errors",
+                    };
+                },
+                projectToAppInfo: async (p: GitProject) => {
+                    const projectClj = await p.findFile("project.clj");
+                    logger.info(`run projectToAppInfo in ${p.baseDir}/${projectClj.path}`);
+                    return {
+                        name: clj.getName(`${p.baseDir}/${projectClj.path}`),
+                        version: clj.getVersion(`${p.baseDir}/${projectClj.path}`),
+                        id: new GitHubRepoRef("owner", "repo"),
+                    };
+                },
+                options: {
+                    env: {
+                        ...process.env,
+                    },
+                },
+            },
+        });
+}
+
 export async function MetajarPreparation(p: GitProject, rwlc: RunWithLogContext): Promise<ExecuteGoalResult> {
     const result = await spawnAndWatch({
-            command: "lein",
-            args: ["with-profile", "metajar", "do", "clean,", "metajar"],
-        },
+        command: "lein",
+        args: ["with-profile", "metajar", "do", "clean,", "metajar"],
+    },
         {
             cwd: p.baseDir,
         },